Refactor: fileify and _manage_imported_files.
