+++ readthedocs/rtd_tests/tests/test_imported_file.py
@@ -3,8 +3,8 @@ import os
 
 from django.test import TestCase
 
+from readthedocs.builds.tasks import _manage_imported_files
 from readthedocs.projects.models import ImportedFile, Project
-from readthedocs.projects.tasks import _manage_imported_files
 
 
 base_dir = os.path.dirname(os.path.dirname(__file__))
